Worked on a number of changes, CimpleG should now be faster and ligher.
Implemented a couple of changes that should make CimpleG faster and more
efficient, particularly when calculating the difference in means and sum
of variances (the fastest will be when both Rfast and matrixStats are
installed).

Added butcher::butcher to the end of the workflow of the complex models, which
should make them much much lighter.

Implemented a way for the user to save the resulting models directly in
the CimpleG function.

Added a verbosity argument to easily control the output logs.

Added a more simple function to do the diffmeans sumvar plots.

Added functions to save and load data in a faster manner.

Renamed the way we call some of the models (eg when running CimpleG,
instead of calling the method "parab_scale", we now call it "CimpleG").

Added more documentation for functions that were lacking it.

Added a few tests (parallel test is not working locally but works on the
cluster).