Mainly changed the algorithm with which we do FS.
Because I had to work on this change I also greatly improved the program
in terms of speed and eficiency. This was mainly done by integrating more
of the data.table framework and less of the dplyr one.
Made a number of changes to the codebase to cope with this change.